Renamed editor to response
