+++ lib/Rpc/Handler/ContextMenuHandler.php
@@ -5,10 +5,10 @@ namespace Phpactor\Rpc\Handler;
 use Phpactor\Rpc\Handler;
 use Phpactor\WorseReflection\Reflector;
 use Phpactor\WorseReflection\Core\SourceCode;
-use Phpactor\Rpc\Editor\EchoResponse;
+use Phpactor\Rpc\Response\EchoResponse;
 use Phpactor\Rpc\Request;
-use Phpactor\Rpc\Editor\InputCallbackResponse;
-use Phpactor\Rpc\Editor\Input\ChoiceInput;
+use Phpactor\Rpc\Response\InputCallbackResponse;
+use Phpactor\Rpc\Response\Input\ChoiceInput;
 use PhpBench\DependencyInjection\Container;
 use Phpactor\Container\RpcExtension;
 use Phpactor\WorseReflection\Core\Offset;
