+++ lib/Rpc/Handler/GotoDefinitionHandler.php
@@ -3,7 +3,7 @@
 namespace Phpactor\Rpc\Handler;
 
 use Phpactor\Rpc\Handler;
-use Phpactor\Rpc\Editor\OpenFileResponse;
+use Phpactor\Rpc\Response\OpenFileResponse;
 use Phpactor\Core\GotoDefinition\GotoDefinition;
 use Phpactor\WorseReflection\Reflector;
 use Phpactor\WorseReflection\Core\SourceCode;
