+++ lib/Rpc/Handler/RenameVariableHandler.php
@@ -3,9 +3,9 @@
 namespace Phpactor\Rpc\Handler;
 
 use Phpactor\CodeTransform\Domain\Refactor\RenameVariable;
-use Phpactor\Rpc\Editor\ReplaceFileSourceResponse;
-use Phpactor\Rpc\Editor\Input\TextInput;
-use Phpactor\Rpc\Editor\Input\ChoiceInput;
+use Phpactor\Rpc\Response\ReplaceFileSourceResponse;
+use Phpactor\Rpc\Response\Input\TextInput;
+use Phpactor\Rpc\Response\Input\ChoiceInput;
 use Phpactor\CodeTransform\Domain\SourceCode;
 
 class RenameVariableHandler extends AbstractHandler
