+++ lib/Rpc/Response/ErrorResponse.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Rpc\Editor;
+namespace Phpactor\Rpc\Response;
 
 use Phpactor\Rpc\Response;
