+++ lib/Rpc/Response/FileReferencesResponse.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Phpactor\Rpc\Editor;
+namespace Phpactor\Rpc\Response;
 
 use Phpactor\Rpc\Response;
-use Phpactor\Rpc\Editor\Reference\FileReferences;
-use Phpactor\Rpc\Editor\Reference\Reference;
+use Phpactor\Rpc\Response\Reference\FileReferences;
+use Phpactor\Rpc\Response\Reference\Reference;
 
 class FileReferencesResponse implements Response
 {
