+++ lib/Rpc/Response/Input/ChoiceInput.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Rpc\Editor\Input;
+namespace Phpactor\Rpc\Response\Input;
 
 // show context menu with "pp"
