+++ lib/Rpc/Response/Input/Input.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Phpactor\Rpc\Editor\Input;
+namespace Phpactor\Rpc\Response\Input;
 
 interface Input
 {
