+++ lib/Rpc/Response/Input/TextInput.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Phpactor\Rpc\Editor\Input;
+namespace Phpactor\Rpc\Response\Input;
+
+use Phpactor\Rpc\Response\Input\Input;
 
 class TextInput implements Input
 {
