+++ tests/Unit/Rpc/Handler/ClassSearchHandlerTest.php
@@ -5,9 +5,9 @@ namespace Phpactor\Tests\Unit\Rpc\Handler;
 use Phpactor\Application\ClassSearch;
 use Phpactor\Rpc\Handler\ClassSearchHandler;
 use Phpactor\Rpc\Handler;
-use Phpactor\Rpc\Editor\EchoResponse;
-use Phpactor\Rpc\Editor\ReturnResponse;
-use Phpactor\Rpc\Editor\ReturnChoiceResponse;
+use Phpactor\Rpc\Response\EchoResponse;
+use Phpactor\Rpc\Response\ReturnResponse;
+use Phpactor\Rpc\Response\ReturnChoiceResponse;
 
 class ClassSearchHandlerTest extends HandlerTestCase
 {
