+++ tests/Unit/Rpc/RequestHandler/LoggingHandlerTest.php
@@ -8,7 +8,7 @@ use Phpactor\Rpc\RequestHandler\ExceptionCatchingHandler;
 use Phpactor\Rpc\Request;
 use Phpactor\Rpc\Response;
 use Prophecy\Prophecy\ObjectProphecy;
-use Phpactor\Rpc\Editor\ErrorResponse;
+use Phpactor\Rpc\Response\ErrorResponse;
 use Psr\Log\LoggerInterface;
 use Phpactor\Rpc\RequestHandler\LoggingHandler;
 use Psr\Log\LogLevel;
